+++ plugins/shn/shn.c
@@ -28,7 +28,6 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <inttypes.h>
 #include <math.h>
 #include "shorten.h"
 #include <deadbeef/deadbeef.h>
@@ -917,7 +916,7 @@ shn_insert (ddb_playlist_t *plt, DB_playItem_t *after, const char *fname) {
     deadbeef->junk_id3v1_read (it, tmp_file->vars.fd);
 
     char s[100];
-    snprintf (s, sizeof (s), "%" PRId64, fsize);
+    snprintf (s, sizeof (s), "%lld", (long long)fsize);
     deadbeef->pl_add_meta (it, ":FILE_SIZE", s);
     snprintf (s, sizeof (s), "%d", tmp_file->wave_header.bits_per_sample);
     deadbeef->pl_add_meta (it, ":BPS", s);
